requirements.txt
Flask
Flask-Bcrypt
Flask-Login
Pillow
Flask-SQLAlchemy
Flask-Migrate
python-dotenv